backend/requirements.txt
fastapi
uvicorn
pymongo
motor
python-dotenv
pydantic
google-generativeai
PyMuPDF
python-multipart

